reorg, update logic for smart metadata class to have a valid? method instead of forcefully validate the instance
